minor renaming & changes
